Refactor GUI manager; add cylinder animation
The GUI method was refactored to a new class to get better organized
code. I've implemented a revolver cylinder animation to show the shop
items in a pretty way.

This cylinder system listens to the rotary detent event to perform the
animation, and uses Bézier curves to define the movement of the hidden
items. It will be used in the game menus, too.